6 - X
Here's an API. Create a web app to call this API with your full name and print the response. This could be extended where we ask you to do something with the response. Like add a text, or capitalise etc.

API link - https://lessonfourapi.tanaypratap.repl.co/translate/yoda.json